Split settings extraction into a traversal core and setting evaluators
settings/extraction.rs interleaved AST traversal, control-flow state,
star-import following, binding state, and per-setting evaluation in
one 1,600-line module, with a second parallel statement walk feeding
branch joins. The extractor is now an entrypoint plus focused modules:
a statement traversal owning control flow, imports, and local
bindings; a binding environment with explicit ambiguous-branch join
semantics; and per-setting evaluators for INSTALLED_APPS and TEMPLATES
that receive a read-only environment value and return plain effects,
with no access to collector internals.

Extraction behavior is unchanged. New characterization tests snapshot
the complete extracted DjangoSettings for representative layouts,
including idioms that currently degrade — split settings importing
names without a star, app lists composed from literal aliases,
duplicate dict keys, unknown **spreads, unsupported mutations, and
aliases assigned in ambiguous branches — so the restructure is pinned
against today's output byte for byte.

The branch-join write scan stays a separate pass on purpose: the main
walk skips statically-false branches, while join conservatism must see
writes in every arm, dead or not.

Star-import extraction results are cached and shared behind Arc
instead of cloning the full binding state on every cache hit and
insert, and ambiguous-branch walking moves branch results instead of
cloning them. One-line setters, defensive re-checks behind guards, and
single-caller pass-through helpers in the rewritten code are
collapsed.
